Migration cleanup
- Remove mono dependency in Travis
- Fix file not found error
  - Filenames are case sensitive in *nix
- Update license headers
- PR Feedback
