+++ benchmarks/Grpc.AspNetCore.Performance/DefaultCoreConfig.cs
@@ -43,7 +43,7 @@ namespace Grpc.AspNetCore.Performance
 
             Add(JitOptimizationsValidator.FailOnError);
 
-            // TODO: Change to out of process and enable server GC when https://github.com/dotnet/BenchmarkDotNet/issues/1023 is fixed
+            // TODO(JamesNK): Change to out of process and enable server GC when https://github.com/dotnet/BenchmarkDotNet/issues/1023 is fixed
             Add(Job.Core
                 .With(CsProjCoreToolchain.From(new NetCoreAppSettings("netcoreapp3.0", null, ".NET Core 3.0")))
                 .With(InProcessToolchain.Instance)
