+++ examples/Clients/Chatter/Program.cs
@@ -1,4 +1,22 @@
-﻿using System;
+﻿#region Copyright notice and license
+
+// Copyright 2019 The gRPC Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#endregion
+
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Common;
@@ -9,8 +27,14 @@ namespace Sample.Clients
 {
     class Program
     {
-        static async Task Main(string[] args)
+        static async Task<int> Main(string[] args)
         {
+            if (args.Length != 1)
+            {
+                Console.WriteLine("No name provided. Usage: dotnet run <name>.");
+                return 1;
+            }
+
             var name = args[0];
             var channel = new Channel("localhost:50051", ClientResources.SslCredentials);
             var client = new Chatter.ChatterClient(channel);
@@ -41,6 +65,8 @@ namespace Sample.Clients
             channel.ShutdownAsync().Wait();
             Console.WriteLine("Press any key to exit...");
             Console.ReadKey();
+
+            return 0;
         }
     }
 }
