+++ examples/Proto/count.proto
@@ -1,5 +1,3 @@
-﻿#region Copyright notice and license
-
 // Copyright 2019 The gRPC Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,8 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#endregion
+syntax = "proto3";
 
-using Grpc.AspNetCore.Performance;
+import "google/protobuf/empty.proto";
+package Count;
 
-[assembly: DefaultCoreConfigAttribute]
+// The counter service definition.
+service Counter {
+  // Get current count
+  rpc IncrementCount (google.protobuf.Empty) returns (CounterReply) {}
+  // Increment count through multiple counts
+  rpc AccumulateCount (stream CounterRequest) returns (CounterReply) {}
+}
+
+// The request message containing the count to increment by
+message CounterRequest {
+  int32 count = 1;
+}
+
+// The response message containing the current count
+message CounterReply {
+  int32 count = 1;
+}
