+++ examples/Proto/greet.proto
@@ -1,5 +1,3 @@
-﻿#region Copyright notice and license
-
 // Copyright 2019 The gRPC Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,8 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#endregion
+syntax = "proto3";
 
-using Grpc.AspNetCore.Performance;
+package Greet;
 
-[assembly: DefaultCoreConfigAttribute]
+// The greeting service definition.
+service Greeter {
+  // Sends a greeting
+  rpc SayHello (HelloRequest) returns (HelloReply) {}
+  rpc SayHellos (HelloRequest) returns (stream HelloReply) {}
+}
+
+// The request message containing the user's name.
+message HelloRequest {
+  string name = 1;
+}
+
+// The response message containing the greetings
+message HelloReply {
+  string message = 1;
+}
