+++ examples/Server/Server.csproj
@@ -1,9 +1,9 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.0</TargetFramework>
-    <LangVersion>latest</LangVersion>
+    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+    <GenerateUserSecretsAttribute>false</GenerateUserSecretsAttribute>
   </PropertyGroup>
 
   <ItemGroup>
@@ -12,14 +12,17 @@
     <!-- 2. Generator metadata: this is the target that's run during the design time build to generate the *.cs files -->
     <!-- 3. None ItemGroup: Remove the *.proto files from None since the generator metadata for those files are ignored -->
     <!-- 4. Content ItemGroup: Include the *.proto files in the Content ItemGroup. This isn't required if the file is included in the project directory. I'm not sure about what's going on with this requirement. -->
-    <ProtoBuf Include="..\..\Proto\count.proto" GrpcServices="Client" Generator="MSBuild:Compile" />
-    <None Remove="@(ProtoBuf)" />
-    <Content Include="@(ProtoBuf)" LinkBase="" />
+    <Protobuf Include="..\Proto\*.proto" GrpcServices="Server" Generator="MSBuild:Compile" />
 
-    <Compile Include="$(MSBuildThisFileDirectory)..\..\Shared\*.cs" />
+    <None Remove="@(Protobuf)" />
+    <Content Include="@(Protobuf)" LinkBase="" />
+
+    <Compile Include="$(MSBuildThisFileDirectory)..\Shared\Resources.cs" />
+
+    <ProjectReference Include="..\..\src\Grpc.AspNetCore.Server\Grpc.AspNetCore.Server.csproj" />
 
     <PackageReference Include="Google.Protobuf" Version="$(GoogleProtobufPackageVersion)" />
-    <PackageReference Include="Grpc.Core" Version="$(GrpcCorePackageVersion)" />
-    <PackageReference Include="Grpc.Tools" Version="$(GrpcToolsPackageVersion)" />
+    <PackageReference Include="Grpc.Tools" Version="$(GrpcToolsPackageVersion)" PrivateAssets="All" />
   </ItemGroup>
+
 </Project>
