+++ examples/Shared/Resources.cs
@@ -1,3 +1,21 @@
+#region Copyright notice and license
+
+// Copyright 2019 The gRPC Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#endregion
+
 using System;
 using System.IO;
 
@@ -5,7 +23,7 @@ namespace Common
 {
     public static class Resources
     {
-        public static string CertDir = Path.Combine(GetSolutionDirectory(), "sample", "Certs");
+        public static string CertDir = Path.Combine(GetSolutionDirectory(), "examples", "Certs");
         public static string ServerPFXPath = Path.Combine(CertDir, "server.pfx");
 
         private static string GetSolutionDirectory()
