--- sample/Proto/chat.proto
@@ -1,15 +1,0 @@
-syntax = "proto3";
-
-package Chat;
-
-// The greeting service definition.
-service Chatter {
-  // Sends a greeting
-  rpc Chat (stream ChatMessage) returns (stream ChatMessage) {}
-}
-
-// The chat message.
-message ChatMessage {
-  string name = 1;
-  string message = 2;
-}
