--- sample/Proto/count.proto
@@ -1,22 +1,0 @@
-syntax = "proto3";
-
-import "google/protobuf/Empty.proto";
-package Count;
-
-// The counter service definition.
-service Counter {
-  // Get current count
-  rpc IncrementCount (google.protobuf.Empty) returns (CounterReply) {}
-  // Increment count through multiple counts
-  rpc AccumulateCount (stream CounterRequest) returns (CounterReply) {}
-}
-
-// The request message containing the count to increment by
-message CounterRequest {
-  int32 count = 1;
-}
-
-// The response message containing the current count
-message CounterReply {
-  int32 count = 1;
-}
