--- sample/Server/IncrementingCounter.cs
@@ -1,12 +1,0 @@
-﻿namespace GRPCServer
-{
-    public class IncrementingCounter
-    {
-        public void Increment(int amount)
-        {
-            Count += amount;
-        }
-
-        public int Count { get; private set; }
-    }
-}
