+++ src/Grpc.AspNetCore.Server/Internal/GrpcMarkerService.cs
@@ -18,6 +18,10 @@
 
 namespace Grpc.AspNetCore.Server.Internal
 {
+    /// <summary>
+    /// A marker class used to determine if all the required gRPC services were added
+    /// to the <see cref="IServiceCollection"/>.
+    /// </summary>
     internal class GrpcMarkerService
     {
     }
