+++ src/Grpc.AspNetCore.Server/Internal/ObjectMethodExecutor/ObjectMethodExecutor.cs
@@ -1,5 +1,20 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+#region Copyright notice and license
+
+// Copyright 2019 The gRPC Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#endregion
 
 using System;
 using System.Collections.Generic;
@@ -117,7 +132,7 @@ namespace Grpc.AspNetCore.Server.Internal
         /// which supplies an awaitable-of-object. This always works, but can incur several extra heap allocations
         /// as compared with using Execute and then using "await" on the result value typecasted to the known
         /// awaitable type. The possible extra heap allocations are for:
-        /// 
+        ///
         /// 1. The custom awaitable (though usually there's a heap allocation for this anyway, since normally
         ///    it's a reference type, and you normally create a new instance per call).
         /// 2. The custom awaiter (whether or not it's a value type, since if it's not, you need a new instance
