+++ src/Grpc.AspNetCore.Server/Properties/AssemblyInfo.cs
